+++ public/app/core/components/AppChrome/TopBar/ProfileButton.tsx
@@ -8,7 +8,6 @@ import { Dropdown, Menu, MenuItem, ToolbarButton, useStyles2 } from '@grafana/ui
 import { contextSrv } from 'app/core/core';
 import { t } from 'app/core/internationalization';
 
-import { ThemeSelectorDrawer } from '../../ThemeSelector/ThemeSelectorDrawer';
 import { enrichWithInteractionTracking } from '../MegaMenu/utils';
 import { NewsContainer } from '../News/NewsDrawer';
 
@@ -22,7 +21,6 @@ export function ProfileButton({ profileNode }: Props) {
   const styles = useStyles2(getStyles);
   const node = enrichWithInteractionTracking(cloneDeep(profileNode), false);
   const [showNewsDrawer, onToggleShowNewsDrawer] = useToggle(false);
-  const [showThemeDrawer, onToggleThemeDrawer] = useToggle(false);
 
   if (!node) {
     return null;
@@ -30,9 +28,6 @@ export function ProfileButton({ profileNode }: Props) {
 
   const renderMenu = () => (
     <TopNavBarMenu node={profileNode}>
-      {config.featureToggles.grafanaconThemes && (
-        <MenuItem icon="palette" onClick={onToggleThemeDrawer} label={t('profile.change-theme', 'Change theme')} />
-      )}
       {config.newsFeedEnabled && (
         <>
           <Menu.Divider />
@@ -57,7 +52,6 @@ export function ProfileButton({ profileNode }: Props) {
         />
       </Dropdown>
       {showNewsDrawer && <NewsContainer onClose={onToggleShowNewsDrawer} />}
-      {showThemeDrawer && <ThemeSelectorDrawer onClose={onToggleThemeDrawer} />}
     </>
   );
 }
