--- public/app/core/components/ThemeSelector/getSelectableThemes.ts
@@ -1,20 +1,0 @@
-import { getBuiltInThemes } from '@grafana/data';
-import { config } from '@grafana/runtime';
-
-export function getSelectableThemes() {
-  const allowedExtraThemes = [];
-
-  if (config.featureToggles.extraThemes) {
-    allowedExtraThemes.push('debug');
-  }
-
-  if (config.featureToggles.grafanaconThemes) {
-    allowedExtraThemes.push('desertbloom');
-    allowedExtraThemes.push('gildedgrove');
-    allowedExtraThemes.push('sapphiredusk');
-    allowedExtraThemes.push('tron');
-    allowedExtraThemes.push('gloom');
-  }
-
-  return getBuiltInThemes(allowedExtraThemes);
-}
