+++ public/locales/pseudo-LOCALE/grafana.json
@@ -2814,8 +2814,7 @@
       "old-password-required": "Øľđ päşşŵőřđ įş řęqūįřęđ",
       "passwords-must-match": "Päşşŵőřđş mūşŧ mäŧčĥ",
       "strong-password-validation-register": "Päşşŵőřđ đőęş ŉőŧ čőmpľy ŵįŧĥ ŧĥę şŧřőŉģ päşşŵőřđ pőľįčy"
-    },
-    "change-theme": "Cĥäŉģę ŧĥęmę"
+    }
   },
   "public-dashboard": {
     "acknowledgment-checkboxes": {
